Mini Refactoring. Bug-Fixing in der Tabelle-Steuerung.

+++ src/main/java/dao/classes/VideoInfoDaoHibernateImp.java
@@ -1,23 +1,27 @@
-package dao;
+package dao.classes;
 
+import dao.interfaces.IVideoInfoDaoHibernate;
 import model.classes.VideoInfo;
 import model.interfaces.IVideoInfo;
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 import util.HibernateUtils;
+import util.validator.ObjectValidator;
 
 import java.util.*;
 
-public class VideoDaoHibernateImp {
+public class VideoInfoDaoHibernateImp implements IVideoInfoDaoHibernate {
 
-    public void saveVideoInfo(final IVideoInfo iVideo) {
+    public void saveVideoInfo(final IVideoInfo iVideoInfo) {
         Session session = null;
 
         try {
             session = HibernateUtils.getSession();
             session.beginTransaction();
-            session.saveOrUpdate(iVideo);
-            session.flush();
+            if(ObjectValidator.validateVideoInfoObject(iVideoInfo)) {
+                session.saveOrUpdate(iVideoInfo);
+                session.flush();
+            }
             session.getTransaction().commit();
         } catch (Exception ex) {
             ex.fillInStackTrace();
@@ -73,8 +77,11 @@ public class VideoDaoHibernateImp {
         try {
             session = HibernateUtils.getSession();
             session.beginTransaction();
-            session.delete(iVideoInfo);
-            session.flush();
+
+            if(ObjectValidator.validateVideoInfoObject(iVideoInfo)) {
+                session.delete(iVideoInfo);
+                session.flush();
+            }
             session.getTransaction().commit();
         } catch (Exception ex) {
             ex.fillInStackTrace();
